requirements.txt
Django==4.2.1
django-cleanup==8.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.0
django-silk==5.3.2
django-filter==23.5
drf-spectacular==0.28.0
drf-yasg==1.21.7
Pillow==10.0.1
uvicorn==0.34.2
openai
tiktoken